refactor(release): move M3U fixture generation out of capture driver
